Changed user Id datatype from Long to int.

+++ Back-end/p2/src/main/java/com/revature/p2/security/JwtAuthenticationFilter.java
@@ -36,7 +36,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 
             if (StringUtils.hasText(jwt) && jwtTokenProvider.validateToken(jwt)){
 
-                Long userId = jwtTokenProvider.getUserIdFromJWT(jwt);
+                Integer userId = jwtTokenProvider.getUserIdFromJWT(jwt) ;
                 User userDetails = customUserDetailsService.loadUserById(userId);
 
                 UsernamePasswordAuthenticationToken authentication = new UsernamePasswordAuthenticationToken(userDetails, null, Collections.emptyList());
